Extract widget load data to loader for sharing with lock screen time line provider
