+++ WordPress/WordPressStatsWidgets/LockScreenWidgets/LockScreenSiteListProvider.swift
@@ -16,6 +16,8 @@ struct LockScreenSiteListProvider<T: HomeWidgetData & LockScreenStatsWidgetData>
         UserDefaults(suiteName: WPAppGroupName)?.object(forKey: AppConfiguration.Widget.Stats.userDefaultsSiteIdKey) as? Int
     }
 
+    private let widgetDataLoader = WidgetDataReader<T>()
+
     func placeholder(in context: Context) -> LockScreenStatsWidgetEntry {
         LockScreenStatsWidgetEntry.siteSelected(placeholderContent, context)
     }
@@ -24,7 +26,7 @@ struct LockScreenSiteListProvider<T: HomeWidgetData & LockScreenStatsWidgetData>
 
         guard let site = configuration.site,
               let siteIdentifier = site.identifier,
-              let widgetData = widgetData(for: siteIdentifier) else {
+              let widgetData = widgetDataLoader.widgetData(for: siteIdentifier) else {
 
             if let siteID = defaultSiteID, let content = T.read()?[siteID] {
                 completion(.siteSelected(content, context))
@@ -53,7 +55,7 @@ struct LockScreenSiteListProvider<T: HomeWidgetData & LockScreenStatsWidgetData>
             return
         }
 
-        guard let widgetData = widgetData(for: configuration, defaultSiteID: defaultSiteID) else {
+        guard let widgetData = widgetDataLoader.widgetData(for: configuration, defaultSiteID: defaultSiteID) else {
             completion(Timeline(entries: [.noData], policy: .never))
             return
         }
@@ -87,34 +89,5 @@ struct LockScreenSiteListProvider<T: HomeWidgetData & LockScreenStatsWidgetData>
                 privateCompletion(.siteSelected(newWidgetData, context))
             }
         }
-    }
-}
-
-
-// MARK: - Widget Data
-
-private extension LockScreenSiteListProvider {
-    /// Returns cached widget data based on the selected site when editing widget and the default site.
-    /// Configuration.site is nil until IntentHandler is initialized.
-    /// Configuration.site can have old value after logging in with a different account. No way to reset configuration when the user logs out.
-    /// Using defaultSiteID if both of these cases.
-    /// - Parameters:
-    ///   - configuration: Configuration of the Widget Site Selection Intent
-    ///   - defaultSiteID: ID of the default site in the account
-    /// - Returns: Widget data
-    func widgetData(for configuration: SelectSiteIntent, defaultSiteID: Int) -> T? {
-
-        /// If configuration.site.identifier has value but there's no widgetData, it means that this identifier comes from previously logged in account
-        return widgetData(for: configuration.site?.identifier ?? String(defaultSiteID))
-        ?? widgetData(for: String(defaultSiteID))
-    }
-
-    func widgetData(for siteID: String) -> T? {
-        /// - TODO: we should not really be needing to do this conversion.  Maybe we can evaluate a better mechanism for site identification.
-        guard let siteID = Int(siteID) else {
-            return nil
-        }
-
-        return T.read()?[siteID]
     }
 }
