+++ WordPress/WordPressStatsWidgets/SiteListProvider.swift
@@ -17,6 +17,8 @@ struct SiteListProvider<T: HomeWidgetData>: IntentTimelineProvider {
         UserDefaults(suiteName: WPAppGroupName)?.object(forKey: AppConfiguration.Widget.Stats.userDefaultsSiteIdKey) as? Int
     }
 
+    private let widgetDataLoader = WidgetDataReader<T>()
+
     func placeholder(in context: Context) -> StatsWidgetEntry {
         StatsWidgetEntry.siteSelected(placeholderContent, context)
     }
@@ -25,7 +27,7 @@ struct SiteListProvider<T: HomeWidgetData>: IntentTimelineProvider {
 
         guard let site = configuration.site,
               let siteIdentifier = site.identifier,
-              let widgetData = widgetData(for: siteIdentifier) else {
+              let widgetData = widgetDataLoader.widgetData(for: siteIdentifier) else {
 
             if let siteID = defaultSiteID, let content = T.read()?[siteID] {
                 completion(.siteSelected(content, context))
@@ -58,7 +60,7 @@ struct SiteListProvider<T: HomeWidgetData>: IntentTimelineProvider {
             return
         }
 
-        guard let widgetData = widgetData(for: configuration, defaultSiteID: defaultSiteID) else {
+        guard let widgetData = widgetDataLoader.widgetData(for: configuration, defaultSiteID: defaultSiteID) else {
             completion(Timeline(entries: [.noData(widgetKind)], policy: .never))
             return
         }
@@ -92,34 +94,6 @@ struct SiteListProvider<T: HomeWidgetData>: IntentTimelineProvider {
                 privateCompletion(.siteSelected(newWidgetData, context))
             }
         }
-    }
-}
-
-// MARK: - Widget Data
-
-private extension SiteListProvider {
-    /// Returns cached widget data based on the selected site when editing widget and the default site.
-    /// Configuration.site is nil until IntentHandler is initialized.
-    /// Configuration.site can have old value after logging in with a different account. No way to reset configuration when the user logs out.
-    /// Using defaultSiteID if both of these cases.
-    /// - Parameters:
-    ///   - configuration: Configuration of the Widget Site Selection Intent
-    ///   - defaultSiteID: ID of the default site in the account
-    /// - Returns: Widget data
-    func widgetData(for configuration: SelectSiteIntent, defaultSiteID: Int) -> T? {
-
-        /// If configuration.site.identifier has value but there's no widgetData, it means that this identifier comes from previously logged in account
-        return widgetData(for: configuration.site?.identifier ?? String(defaultSiteID))
-            ?? widgetData(for: String(defaultSiteID))
-    }
-
-    func widgetData(for siteID: String) -> T? {
-        /// - TODO: we should not really be needing to do this conversion.  Maybe we can evaluate a better mechanism for site identification.
-        guard let siteID = Int(siteID) else {
-            return nil
-        }
-
-        return T.read()?[siteID]
     }
 }
